END:: chapter 3-02

+++ app/Http/Controllers/Tweet/DeleteController.php
@@ -3,9 +3,13 @@
 namespace App\Http\Controllers\Tweet;
 
 use App\Http\Controllers\Controller;
+use App\Models\Tweet;
+use App\Services\TweetService;
+
 use Illuminate\Http\Request;
 
-use App\Models\Tweet;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
+
 
 class DeleteController extends Controller
 {
@@ -16,6 +20,9 @@ class DeleteController extends Controller
     {
         //
         $tweetId = (int) $request->route('tweetId');
+        if (!$tweetService -> checkOwnTweet($request -> user() -> id, $tweetId)){
+            throw new AccessDeniedHttpException();
+        }
         //$tweet = Tweet::where('id', $tweetId)->firstOrFail();
         //$tweet->delete();
         $tweet = Tweet::destroy($tweetId);
